frontend update global vue api usage for vue3
