+++ packages/frontend/src/router.ts
@@ -1,5 +1,8 @@
-import Vue from 'vue';
-import Router from 'vue-router';
+import {
+  createRouter,
+  createWebHistory,
+  type NavigationGuard,
+} from 'vue-router';
 import NotFound from './views/NotFound.vue';
 import PostCreate from './views/PostCreate.vue';
 import PostsCategory from './views/PostsCategory.vue';
@@ -9,9 +12,7 @@ import UserPage from './views/UserPage.vue';
 import UserSettings from './views/UserSettings.vue';
 import store from '@/store/index';
 
-Vue.use(Router);
-
-const authGuard = async (to, from, next) => {
+const authGuard: NavigationGuard = async (to, from, next) => {
   // TODO: Why check on each move if the user is logged in
   await store.dispatch('userGetAuthState');
 
@@ -28,14 +29,13 @@ const authGuard = async (to, from, next) => {
   next();
 };
 
-const router = new Router({
-  mode: 'history',
-  base: '/',
+export const router = createRouter({
+  history: createWebHistory('/'),
   scrollBehavior(to) {
     if (to.hash) {
       return { selector: to.hash };
     }
-    return { x: 0, y: 0 };
+    return { x: 0, top: 0 };
   },
   routes: [
     {
@@ -165,7 +165,7 @@ const router = new Router({
 });
 
 router.beforeEach((to, from, next) => {
-  // set title
+  // Setting page title
   const ending = ' | Smiler';
   const titleTo = to.meta.title;
   const titleParams = to.meta.titleParam;
@@ -178,5 +178,3 @@ router.beforeEach((to, from, next) => {
 
   next();
 });
-
-export default router;
