+++ packages/frontend/src/store/index.ts
@@ -1,13 +1,12 @@
-import Vue from 'vue';
-import Vuex from 'vuex';
+import { createStore } from 'vuex';
 import notifications from './notifications';
 import user from './user';
 
-Vue.use(Vuex);
-
-export default new Vuex.Store({
+const store = createStore({
   modules: {
     user,
     notifications,
   },
 });
+
+export default store;
