+++ packages/frontend/src/utils/check-can-edit-comment.ts
@@ -1,0 +1,19 @@
+import store from '@/store';
+import consts from '@/const/const';
+
+export function checkCanEditComment(comment: any): boolean {
+  const currentId = store.state.user.id;
+
+  if (!currentId || !comment.author || comment.author.id !== currentId) {
+    return false;
+  }
+
+  const dateNow = new Date().getTime();
+  const dateCreated = new Date(comment.createdAt).getTime();
+
+  if (dateNow - dateCreated > consts.COMMENT_TIME_TO_UPDATE) {
+    return false;
+  }
+
+  return true;
+}
