+++ packages/frontend/src/utils/format-from-now.ts
@@ -1,0 +1,12 @@
+import { formatDistanceToNowStrict } from 'date-fns';
+
+export function formatFromNow(date: Date | string | number): string {
+  if (!date) {
+    return '';
+  }
+
+  return formatDistanceToNowStrict(date, {
+    addSuffix: true,
+    roundingMethod: 'trunc',
+  });
+}
