+++ packages/frontend/src/utils/is-desktop.ts
@@ -1,0 +1,5 @@
+import { isMobile } from './is-mobile';
+
+export function isDesktop(): boolean {
+  return !isMobile();
+}
