+++ packages/frontend/src/utils/is-mobile.ts
@@ -1,0 +1,7 @@
+export function isMobile(): boolean {
+  if (window.matchMedia('(max-device-width: 600px)').matches) {
+    return true;
+  }
+
+  return false;
+}
