+++ packages/frontend/src/utils/resolve-avatar.ts
@@ -1,0 +1,9 @@
+import defaultAvatar from '@/assets/neutral_avatar.png';
+
+export function resolveAvatar(path: string): string {
+  if (path) {
+    return path;
+  }
+
+  return defaultAvatar;
+}
