+++ packages/frontend/src/utils/resolve-image-error.ts
@@ -1,0 +1,5 @@
+import postNoImage from '@/assets/post_no_image.svg';
+
+export function resolveImageError(e: Event): void {
+  e.target.src = postNoImage;
+}
