+++ packages/frontend/src/utils/resolve-image.ts
@@ -1,0 +1,13 @@
+import config from '@/config/config';
+
+export function resolveImage(path: string): string {
+  if (
+    /https?:\/\/(www\.)?[-a-zA-Z0-9@:%._+~#=]{2,256}\.[a-z]{2,4}\b([-a-zA-Z0-9@:%_+.~#?&//=]*)/.test(
+      path,
+    )
+  ) {
+    return path;
+  }
+
+  return config.VUE_APP_API_URL + path;
+}
